code/preprocessing/requirements.txt
kafka-python
elasticsearch
pandas
numpy
colorlog
spacy
bs4
requests
fuzzywuzzy
langdetect
python-Levenshtein
pycountry
python-restcountries
lxml
flask
flask_cors
flask_swagger_ui
python-whois
tld
tldextract
unidecode
pytz
pillow
dotmap